14-회원 탈퇴 구현/회원가입시 닉네임 추가 및 회원탈퇴 시 암호화된 password 검증 후 true이면 삭제

+++ controllers/users/deleteUser.js
@@ -1,7 +1,17 @@
 const { user } = require('../../models');
+const bcrypt = require('bcrypt');
 
 module.exports = async (req, res) => {
-    await user.destroy({ where: { email: req.body.email } });
-    console.log("deleteUser Complete")
-    res.status(200).send({ data: null, message: 'deleteUser complete' });
+
+    // 비밀번호를 검증
+    const password = await bcrypt.compare(req.body.password, (await user.findOne({ where: {email: req.body.email}})).dataValues.password);
+    
+    if(password) {
+        await user.destroy({where: {email: req.body.email}});
+        res.status(200).send({ data: null, message: 'deleteUser complete'})
+    }
+    else {
+        res.status(400).send({ data: null, message: 'unComplete deleteUser'})
+    }
+    
 }
